refactor: split form config into seperate components and files
