+++ frontend/src/app/admin/form-config/page.tsx
@@ -1,61 +1,15 @@
 'use client';
 
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { ArrowLeft } from 'lucide-react';
 import Link from 'next/link';
 import Navbar from '@/components/Navbar';
-import { DragDropContext, Droppable, Draggable } from '@hello-pangea/dnd';
-
-// Types for API responses
-interface APIResponse<T> {
-  data: T;
-  error?: string;
-}
-
-interface FormQuestion {
-  id: string;
-  text: string;
-  type: 'text' | 'single select' | 'multi select';
-  required: boolean;
-  options?: string[];
-  order: number; // Added for backend ordering
-}
-
-// Mock data - Replace with API call later
-const INITIAL_QUESTIONS: FormQuestion[] = [
-  {
-    id: '1',
-    text: 'Grant Scheme',
-    type: 'single select',
-    required: true,
-    options: ['NHMRC', 'MRFF', 'ARC', 'Other'],
-    order: 0
-  },
-  {
-    id: '2',
-    text: 'Involves MRI',
-    type: 'single select',
-    required: true,
-    options: ['Yes', 'No', 'Other'],
-    order: 1
-  },
-  {
-    id: '3',
-    text: 'Type of Query',
-    type: 'single select',
-    required: true,
-    options: ['Contractual clause review', 'Support with negotiations', 'Advice on agreement type', 'Compliance advice', 'Other'],
-    order: 2
-  },
-  {
-    id: '4',
-    text: 'Mark as Urgent',
-    type: 'single select',
-    required: false,
-    options: ['Yes', 'No'],
-    order: 3
-  }
-];
+import { QuestionList } from '@/components/QuestionList';
+import { 
+  FormQuestion, 
+  fetchMockQuestions,
+  reorderMockQuestions 
+} from './__mocks__/questions';
 
 // API endpoints - Update when backend is implemented
 const API_ENDPOINTS = {
@@ -81,16 +35,11 @@ export default function FormConfigPage() {
 
   // Fetch questions on component mount
   useEffect(() => {
-    const fetchQuestions = async () => {
+    const loadQuestions = async () => {
       try {
         setIsLoading(true);
-        // TODO: Replace with actual API call
-        // const response = await fetch(API_ENDPOINTS.QUESTIONS);
-        // const data: APIResponse<FormQuestion[]> = await response.json();
-        
-        // Temporary: Use mock data
-        await new Promise(resolve => setTimeout(resolve, 500)); // Simulate API delay
-        setQuestions(INITIAL_QUESTIONS);
+        const response = await fetchMockQuestions();
+        setQuestions(response.data);
       } catch (err) {
         setError('Failed to load questions');
       } finally {
@@ -98,7 +47,7 @@ export default function FormConfigPage() {
       }
     };
 
-    fetchQuestions();
+    loadQuestions();
   }, []);
 
   const handleDragEnd = async (result: any) => {
@@ -196,107 +145,27 @@ export default function FormConfigPage() {
               Manage Form Questions
             </h2>
 
-            {/* Conditional rendering based on loading and error states */}
             {isLoading ? (
-              <div className="text-gray-500">Loading configuration...</div>
+              <div className="py-8">
+                <div className="flex justify-center">
+                  <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-900" />
+                </div>
+                <p className="text-center text-gray-500 mt-4">Loading questions...</p>
+              </div>
             ) : error ? (
-              <div className="text-red-600">{error}</div>
+              <div className="py-8">
+                <p className="text-center text-red-600">{error}</p>
+              </div>
             ) : (
               <>
-                <div className="max-h-[360px] overflow-y-auto">
-                  <DragDropContext onDragEnd={handleDragEnd}>
-                    <Droppable droppableId="questions">
-                      {(provided) => (
-                        <div 
-                          {...provided.droppableProps} 
-                          ref={provided.innerRef}
-                          className="space-y-4"
-                        >
-                          {questions.map((question, index) => (
-                            <Draggable 
-                              key={question.id} 
-                              draggableId={question.id} 
-                              index={index}
-                            >
-                              {(provided) => (
-                                <div
-                                  ref={provided.innerRef}
-                                  {...provided.draggableProps}
-                                  {...provided.dragHandleProps}
-                                  className="bg-white border border-gray-200 rounded-lg p-4 mb-4"
-                                >
-                                  <div className="flex items-start justify-between">
-                                    <div className="flex-1">
-                                      <h3 className="font-medium text-gray-900">
-                                        {question.text}
-                                      </h3>
-                                      <div className="flex flex-wrap items-center gap-2 mt-1">
-                                        <span className="px-2 py-0.5 bg-gray-100 text-sm text-gray-600 rounded">
-                                          {question.type}
-                                        </span>
-                                        {question.required && (
-                                          <span className="px-2 py-0.5 bg-gray-100 text-sm text-gray-600 rounded">
-                                            Required
-                                          </span>
-                                        )}
-                                        {question.dependsOn && (
-                                          <span className="text-sm text-gray-500">
-                                            Shows when: {question.dependsOn.value}
-                                          </span>
-                                        )}
-                                      </div>
-                                      {question.options && (
-                                        <p className="text-sm text-gray-500 mt-1">
-                                          Options: {question.options.join(', ')}
-                                        </p>
-                                      )}
-                                    </div>
-                                    <div className="flex items-center gap-2">
-                                      <button
-                                        onClick={() => handleEdit(question)}
-                                        className="p-2 hover:bg-gray-50 rounded-lg"
-                                        title="Edit question"
-                                      >
-                                        <svg className="w-4 h-4 text-gray-600" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15.232 5.232l3.536 3.536m-2.036-5.036a2.5 2.5 0 113.536 3.536L6.5 21.036H3v-3.572L16.732 3.732z" />
-                                        </svg>
-                                      </button>
-                                      <button
-                                        onClick={() => handleDelete(question.id)}
-                                        className="p-2 hover:bg-gray-50 rounded-lg"
-                                        title="Delete question"
-                                      >
-                                        <svg className="w-4 h-4 text-gray-600" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                                        </svg>
-                                      </button>
-                                    </div>
-                                  </div>
-                                  
-                                  {/* Render sub-questions */}
-                                  {question.subQuestions && (
-                                    <div className="ml-6 mt-4 border-l-2 border-gray-200 pl-4">
-                                      {question.subQuestions.map(subQuestion => (
-                                        // Recursively render sub-questions using the same component
-                                        <QuestionCard 
-                                          key={subQuestion.id}
-                                          question={subQuestion}
-                                          onEdit={handleEdit}
-                                          onDelete={handleDelete}
-                                        />
-                                      ))}
-                                    </div>
-                                  )}
-                                </div>
-                              )}
-                            </Draggable>
-                          ))}
-                          {provided.placeholder}
-                        </div>
-                      )}
-                    </Droppable>
-                  </DragDropContext>
-                </div>
+                <QuestionList 
+                  questions={questions}
+                  onDragEnd={handleDragEnd}
+                  // TODO: Implement edit and delete handlers
+                  // onEdit={handleEdit}
+                  // onDelete={handleDelete}
+                />
+                
                 {/* Keep Add New Question button outside scroll area */}
                 <div className="mt-6 pt-6 border-t border-gray-200">
                   <button
